refactor: extracted aggregation pipeline stages
